+++ src/regions/regions.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, ConflictException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateRegionDto } from './dto/create-region.dto';
 import { UpdateRegionDto } from './dto/update-region.dto';
@@ -33,7 +33,30 @@ export class RegionsService {
     }
 
     async remove(id: number) {
-        await this.findOne(id);
+        const region = await this.findOne(id);
+
+        // Check for child entities that would be orphaned
+        const areaCount = await this.prisma.area.count({
+            where: { regionId: id },
+        });
+
+        const retailerCount = await this.prisma.retailer.count({
+            where: { regionId: id },
+        });
+
+        if (areaCount > 0 || retailerCount > 0) {
+            const errors = [];
+            if (areaCount > 0) {
+                errors.push(`${areaCount} area(s) assigned to this region`);
+            }
+            if (retailerCount > 0) {
+                errors.push(`${retailerCount} retailer(s) assigned to this region`);
+            }
+            throw new ConflictException(
+                `Cannot delete region "${region.name}". Please remove the following dependencies first: ${errors.join(', ')}.`
+            );
+        }
+
         return this.prisma.region.delete({ where: { id } });
     }
 }
